+++ src/components/SearchForm.js
@@ -1,4 +1,6 @@
 import React from "react";
+import "../css/SearchForm.css";
+import "../css/main.css";
 
 class SearchForm extends React.Component {
   render() {
