--- src/components/general-styles/main.css
@@ -1,14 +1,0 @@
-@import './common/styles/palette.css';
-@import './common/styles/article.css';
-@import './common/styles/colorVariables.css';
-@import './common/styles/general.css';
-@import './common/styles/header.css';
-@import './common/styles/meetTeam.css';
-@import './common/styles/splash.css';
-@import './common/styles/aura_inputs.css';
-@import './components/search-form/SearchForm.css';
-@import './components/result-card/ResultCard.css';
-
-:root {
-  font-size: 18px;
-}
