+++ src/css/general.css
@@ -9,7 +9,7 @@
 }
 
 body {
-  background-image: url(../../assets/img/trinityBackgroundWhite.png);
+  background-image: url(../assets/img/trinityBackgroundWhite.png);
   margin: 0;
 }
 
@@ -41,9 +41,11 @@ h2 {
   display: block;
   text-align: center;
   box-shadow: 0px 0px 3px 2px rgba(0, 0, 0, 0.2);
-  background: radial-gradient(circle,
+  background: radial-gradient(
+    circle,
     rgba(39, 39, 39, 0.85),
-    rgba(0, 0, 0, 0.27) 70%);
+    rgba(0, 0, 0, 0.27) 70%
+  );
 }
 
 section img ul h3 {
