+++ src/css/header.css
@@ -1,4 +1,4 @@
-@import "../general-styles/colorVariables.css";
+@import "./colorVariables.css";
 
 /*Header*/
 header {
