+++ src/temp/main.js
